Removed the InstalledFeatureDependency in favour of InstalledFeatureRef (that's why this is a v1alpha1). And broke the InstalledFeature controller into several files to be better handleable than a multi-hundred line file.
