+++ config/samples/features_v1alpha1_installedfeaturegroup.yaml
@@ -13,4 +13,6 @@ status:
   message: just installing
   features:
     - name: k8s-feature-library
+      namespace: default
     - name: k8s-feature-library-operator
+      namespace: default
