+++ go.mod
@@ -5,6 +5,7 @@ go 1.13
 require (
 	github.com/go-logr/logr v0.1.0
 	github.com/golang/mock v1.2.0
+	github.com/google/uuid v1.1.1
 	github.com/onsi/ginkgo v1.12.1
 	github.com/onsi/gomega v1.10.1
 	github.com/ory/go-acc v0.2.6 // indirect
